Fix evaluation mode + variable assignment bug
AnalyzeEmbeddedExpression skips ahead to the first variable assignment
it can find, ignoring any other evaluation mode blocks along the way.
This works fine for dealing with the global variable declarations, but
throws a wrench in the works when there's a later, unrelated variable
assignment after an evaluation mode block at the same level in the
container.

Since (at least) temp variable assignments happen outside of an
evaluation context, abort skipping ahead if we enter the start of a new
one. I am not 100% sure this is correct, but it seems to work