Stop ContainerCursor.Current OOB access
encountered thanks to some code in the wild shaped like this handcrafted
testcase. the noteworthy thing here is that the variable assignment
happens inside the eval context, which then immediately ends. i'm not
actually sure how to get ink to *generate* code shaped like that
(my current suspicion is that it was generated by an older version), so
i'm not 100% sure if the provided ink is actually what generated this case